Address review: support signed Mongo cursor limits
https://github.com/ClickHouse/ClickHouse/pull/68493

+++ src/Parsers/Mongo/Metadata.h
@@ -46,8 +46,8 @@ public:
         std::string database_name_,
         std::string collection_name_,
         QueryType query_type_,
-        std::optional<int> limit_,
-        std::optional<int> offset_,
+        std::optional<Int64> limit_,
+        std::optional<Int64> offset_,
         std::optional<std::string> order_by_);
 
     /// Empty when the query does not name a database, which means the current one.
@@ -57,9 +57,9 @@ public:
 
     QueryType getQueryType() const { return query_type; }
 
-    std::optional<size_t> getLimit() const { return limit; }
+    std::optional<UInt64> getLimit() const { return limit; }
 
-    std::optional<size_t> getOffset() const { return offset; }
+    std::optional<UInt64> getOffset() const { return offset; }
 
     std::optional<std::string> getOrderBy() const { return order_by; }
 
@@ -71,8 +71,8 @@ private:
     std::string database_name;
     std::string collection_name;
     QueryType query_type;
-    std::optional<size_t> limit;
-    std::optional<size_t> offset;
+    std::optional<UInt64> limit;
+    std::optional<UInt64> offset;
     std::optional<std::string> order_by;
     rapidjson::Document::AllocatorType allocator;
 };
